refactor hf dataloader
