+++ internlm/data/build_dataloader.py
@@ -1,17 +1,14 @@
 # Copyright (c) InternLM. All rights reserved.
 from functools import partial
-import sys
-import datasets
 
-import torch
 import torch.distributed as dist
-from transformers import AutoTokenizer
-from datasets.distributed import split_dataset_by_node
-from torch.utils.data import ConcatDataset, DataLoader, Dataset
+from torch.utils.data import ConcatDataset, DataLoader
 
 from internlm.core.context import ParallelMode
 from internlm.core.context import global_context as gpc
 from internlm.data.streaming.batch_sampler import StreamingStaticBatchSampler
+from internlm.data.streaming.collaters import hf_collate_fn
+from internlm.data.streaming.dataset import HuggingFaceStreamingDataset
 from internlm.data.tokenized.batch_sampler import (
     StaticBatchSampler,
     get_dpsampler_dataloader,
@@ -113,72 +110,12 @@ def get_tokenized_valid_loader_items(data_cfg):
 
     return valid_ds, valid_collate_fn
 
-def create_hf_dataloader(data_cfg, split='train'):
-    def collate_fn(batch, micro_num, micro_bsz, seq_len):
-        input_ids_list = []
-        attention_mask_list = []
-        labels_list = []
-        for b in batch:
-            attention_mask = b['attention_mask']
-            input_ids = b['input_ids']
-            input_ids = torch.abs(input_ids * attention_mask)
-            input_ids = torch.nn.functional.pad(input_ids, (0, seq_len - len(input_ids)), mode='constant', value=0)
-            attention_mask = torch.nn.functional.pad(attention_mask, (0, seq_len - len(attention_mask)), mode='constant', value=0)
-            label = torch.tensor([w if w > 0 else -100 for w in input_ids.tolist()][1:]+[-100])
-            input_ids_list.append(input_ids)
-            attention_mask_list.append(attention_mask)
-            labels_list.append(label)
-        input_ids = torch.stack(input_ids_list)
-        attention_mask = torch.stack(attention_mask_list)
-        labels = torch.stack(labels_list)
-        return {"input_ids": input_ids, "attention_mask": attention_mask, "type_ids": torch.zeros(micro_num, micro_bsz, seq_len, dtype=torch.int64)}, labels
 
-    train_dataset = HuggingFaceStreamingDataset(data_cfg.train_folder, data_cfg.tokenizer_path, data_cfg.seq_len, split)
-    train_batch_sampler = StreamingStaticBatchSampler(batch_size = data_cfg.micro_num * data_cfg.micro_bsz, rampup_batch_size = data_cfg.rampup_batch_size)
-    train_dl = DataLoader(
-        dataset=train_dataset,
-        batch_sampler=train_batch_sampler,
-        num_workers=data_cfg.get("num_worker", 4),
-        pin_memory=True,
-        collate_fn=partial(collate_fn, micro_num=data_cfg.micro_num, micro_bsz=data_cfg.micro_bsz, seq_len=data_cfg.seq_len),
-        persistent_workers=data_cfg.get("num_worker", 4) > 0,
-    )
-    return train_dl
-
-class HuggingFaceStreamingDataset(Dataset):
-    def __init__(self, dataset_name, tokenizer_name, model_max_length, split='train', buffer_size=1000):
-        self.dataset = datasets.load_dataset(dataset_name, split=split, streaming=True)
-        self.dataset = split_dataset_by_node(self.dataset, rank=gpc.get_local_rank(ParallelMode.DATA), world_size=gpc.get_world_size(ParallelMode.DATA))
-        self.buffer_size = buffer_size
-        self.senior_iterator = iter(self)
-
-        self.tokenizer = AutoTokenizer.from_pretrained(tokenizer_name, trust_remote_code=True)
-        self.tokenizer.padding_side = "right"
-        self.tokenizer.truncation_side = "right"
-        self.tokenizer.model_max_length = model_max_length
-
-    def __iter__(self):
-        buffer = []
-        for sample in self.dataset:
-            buffer.append(sample)
-            if len(buffer) >= self.buffer_size:
-                yield from self._tokenize(buffer)
-                buffer = []
-
-        if buffer:
-            yield from self._tokenize(buffer)
-    
-    def __len__(self):
-        return sys.maxsize
-    
-    def _tokenize(self, samples):
-        texts = [sample['text'] for sample in samples]
-        tokenized_outputs = self.tokenizer(texts, padding=True, truncation=True, return_tensors='pt')
-        for i in range(len(samples)):
-            yield {key: tokenized_outputs[key][i] for key in tokenized_outputs}
-
-    def __getitem__(self, _):
-        return next(self.senior_iterator)
+def get_hf_train_loader_items(data_cfg):
+    train_ds = HuggingFaceStreamingDataset(data_cfg.train_folder, data_cfg.tokenizer_path, data_cfg.seq_len)
+    train_sampler = StreamingStaticBatchSampler(batch_size = data_cfg.micro_num * data_cfg.micro_bsz, rampup_batch_size = data_cfg.rampup_batch_size)
+    train_collate_fn = partial(hf_collate_fn, micro_num=data_cfg.micro_num, micro_bsz=data_cfg.micro_bsz, seq_len=data_cfg.seq_len)
+    return train_ds, train_sampler, train_collate_fn
 
 
 def build_train_loader_with_data_type():
@@ -189,15 +126,14 @@ def build_train_loader_with_data_type():
     """
     data_cfg = gpc.config.data
 
-    if data_cfg.type == "hf":
-        train_dl = create_hf_dataloader(data_cfg)
-        return train_dl, ["en"]
-
     train_folder = data_cfg.get("train_folder", None)
-    dataset_types = list(get_dataset_type_ids_map(train_folder).keys()) if train_folder else ["en", "cn", "code"]
 
     if data_cfg.type == "tokenized":
         train_ds, train_sampler, train_collate_fn = get_tokenized_train_loader_items(data_cfg)
+        dataset_types = list(get_dataset_type_ids_map(train_folder).keys()) if train_folder else ["en", "cn", "code"]
+    elif data_cfg.type == "hf":
+        train_ds, train_sampler, train_collate_fn = get_hf_train_loader_items(data_cfg)
+        dataset_types = ["en"]
     else:
         raise ValueError(f"dataset type {data_cfg.type} is not supported")
