+++ internlm/data/streaming/__init__.py
@@ -1,5 +1,9 @@
 from .batch_sampler import StreamingStaticBatchSampler
+from .collaters import hf_collate_fn
+from .dataset import HuggingFaceStreamingDataset
 
 __all__ = [
     "StreamingStaticBatchSampler",
+    "hf_collate_fn",
+    "HuggingFaceStreamingDataset",
 ]
